requirements.txt
# --------- pytorch --------- #
torch==2.0.1
torchvision>=0.15.0
lightning>=2.0.0
torchmetrics>=0.11.4

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0


# --------- others --------- #
rootutils       # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
einops
# sh            # for running bash commands in some tests (linux/macos only)
datasets==2.14.6
entmax==1.1
hydra-core==1.3.2
ipython==8.15.0
jsonlines==3.1.0
numpy==1.25.2
omegaconf==2.3.0
pandas==2.0.3
pickleshare==0.7.5
pytorch-lightning==2.0.7
pyyaml==6.0.1
scikit-learn==1.3.0
scipy==1.11.2
tqdm==4.66.1
transformers==4.32.0
wandb==0.15.8
sentencepiece==0.2.0
# pynini==2.1.5
